Delete sentence from IDCT post about entropy coding of repeated values
Upon further consideration, this paragraph is not very correct. Yes,
_generally_ entropy coding works well when there are many repeated
values in your data. But in JPEG scan sections, Huffman coding is not
actually applied to the non-zero coefficient values, but rather to
1) the lengths of runs of zero coefficients, and 2) the number of
significant bits in each non-zero coefficient.

Arithmetic-coded JPEGs are a different beast, and for those, this
paragraph may be more on point. Anyways, I think it's better to just
drop it.