Tweak animated trees in Huffman post for clarity
Move the text for the node 'weight' up a bit so they don't overlap the
lines linking nodes together as much. Also, the animation which shows
the process of building an optimal tree becomes too crowded if a long
input string is entered... so limit the length of the input.