So much for my responsive design
Sigh. I had wanted the function plots in my Nelder-Mead explainer to
be square on narrow screens, but rectangular on wide screens. The idea
was that the SVG function plot would stretch as necessary, but that
the points would always be circular; to make this happen, I used
zero-length 'line' elements for the circles, with a round endcap.

The plan worked beautifully in desktop and mobile Chrome. But not
in Safari browser. Oh, what could have been, if only...

A stronger man than I would battle onward and fufill his vision
at all costs. But I am just going to give up and make all the
function plots square.